Move around package names to fix package name issues
