+++ desktop/src/main/java/nl/tudelft/contextproject/desktop/ExitListener.java
@@ -1,4 +1,4 @@
-package nl.tudelft.contextproject.java;
+package nl.tudelft.contextproject.desktop;
 
 import com.badlogic.gdx.LifecycleListener;
