+++ image-processing/src/main/java/nl/tudelft/contextproject/ImageProcessing.java
@@ -1,4 +1,4 @@
-package nl.tudelft.context.project.imageprocessing;
+package nl.tudelft.contextproject;
 
 public class ImageProcessing {
     public ImageProcessing() {
